Partially update to post 1.0 libs and syntax

+++ src/hash.rs
@@ -1,6 +1,6 @@
 use dct::{dct_2d, crop_dct};
 
-use image::{GenericImage, DynamicImage, 
+use image::{GenericImage, DynamicImage,
     ImageBuf, Luma, Pixel, FilterType, Nearest, Rgba};
 use image::imageops::{grayscale, resize};
 
@@ -10,7 +10,7 @@ use std::collections::Bitv;
 
 const FILTER_TYPE: FilterType = Nearest;
 
-#[deriving(PartialEq, Eq, Hash, Show, Clone)]
+#[derive(PartialEq, Eq, Hash, Show, Clone)]
 pub struct ImageHash {
     size: u32,
     bitv: Bitv,
@@ -18,8 +18,8 @@ pub struct ImageHash {
 
 impl ImageHash {
 
-    pub fn dist(&self, other: &ImageHash) -> uint {
-        assert!(self.bitv.len() == other.bitv.len(), 
+    pub fn dist(&self, other: &ImageHash) -> usize {
+        assert!(self.bitv.len() == other.bitv.len(),
                 "ImageHashes must be the same length for proper comparison!");
 
         self.bitv.iter().zip(other.bitv.iter())
@@ -28,27 +28,27 @@ impl ImageHash {
 
     pub fn dist_ratio(&self, other: &ImageHash) -> f32 {
         self.dist(other) as f32 / self.size as f32
-    }    
+    }
 
-    
+
     fn fast_hash<Img: GenericImage<Rgba<u8>>>(img: &Img, hash_size: u32) -> Bitv {
         let temp = square_resize_and_gray(img, hash_size);
 
         let hash_values: Vec<u8> = temp.pixels().map(|(_, _, x)| x.channel())
             .collect();
 
-        let hash_sq = (hash_size * hash_size) as uint;
+        let hash_sq = (hash_size * hash_size) as usize;
 
-        let mean = hash_values.iter().fold(0u, |b, &a| a as uint + b) 
+        let mean = hash_values.iter().fold(0u, |b, &a| a as usize + b)
             / hash_sq;
 
-        hash_values.into_iter().map(|x| x as uint >= mean).collect()
+        hash_values.into_iter().map(|x| x as usize >= mean).collect()
     }
 
     fn dct_hash<Img: GenericImage<Rgba<u8>>>(img: &Img, hash_size: u32) -> Bitv {
         let large_size = hash_size * 4;
 
-        // We take a bigger resize than fast_hash, 
+        // We take a bigger resize than fast_hash,
         // then we only take the lowest corner of the DCT
         let temp = square_resize_and_gray(img, large_size);
 
@@ -57,27 +57,27 @@ impl ImageHash {
             .map(|(_, _, x)| x.channel() as f64).collect();
 
         let dct = dct_2d(hash_values.as_slice(),
-            large_size as uint, large_size as uint);
+            large_size as usize, large_size as usize);
 
-        let original = (large_size as uint, large_size as uint);
-        let new = (hash_size as uint, hash_size as uint);
+        let original = (large_size as usize, large_size as usize);
+        let new = (hash_size as usize, hash_size as usize);
 
         let cropped_dct = crop_dct(dct, original, new);
 
-        let mean = cropped_dct.iter().fold(0f64, |b, &a| a + b) 
+        let mean = cropped_dct.iter().fold(0f64, |b, &a| a + b)
             / (hash_size * hash_size) as f64;
 
         cropped_dct.into_iter().map(|x| x >= mean).collect()
-    }    
+    }
 
     pub fn hash<Img: GenericImage<Rgba<u8>>>(img: &Img, hash_size: u32, fast: bool) -> ImageHash {
-        let hash = if fast { 
-            ImageHash::fast_hash(img, hash_size)   
-        } else { 
-            ImageHash::dct_hash(img, hash_size)             
+        let hash = if fast {
+            ImageHash::fast_hash(img, hash_size)
+        } else {
+            ImageHash::dct_hash(img, hash_size)
         };
 
-        assert!((hash_size * hash_size) as uint == hash.len());
+        assert!((hash_size * hash_size) as usize == hash.len());
 
         ImageHash {
             size: hash_size * hash_size,
